+++ src/main/java/net/kyori/registry/DefaultedRegistryGetter.java
@@ -21,54 +21,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.registry.api.registry;
+package net.kyori.registry;
 
 import org.checkerframework.checker.nullness.qual.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
 
-import java.util.Iterator;
-import java.util.Set;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
+public interface DefaultedRegistryGetter<K, V> extends RegistryGetter<K, V> {
+  @NonNull K defaultKey();
 
-public interface RegistryGetter<K, V> extends Iterable<V> {
-  /**
-   * Gets the value for {@code key}.
-   *
-   * @param key the key
-   * @return the value
-   */
-  @Nullable V get(final @NonNull K key);
-
-  /**
-   * Gets the key for {@code value}.
-   *
-   * @param value the value
-   * @return the key
-   */
-  @Nullable K key(final @NonNull V value);
-
-  /**
-   * Gets a set of the keys contained in this registry.
-   *
-   * @return a set of the keys contained in this registry
-   */
-  @NonNull Set<K> keySet();
-
-  /**
-   * Creates an unmodifiable iterator of values.
-   *
-   * @return an unmodifiable iterator
-   */
-  @Override
-  @NonNull Iterator<V> iterator();
-
-  /**
-   * Creates a stream of values.
-   *
-   * @return a stream
-   */
-  default @NonNull Stream<V> stream() {
-    return StreamSupport.stream(this.spliterator(), false);
-  }
+  @NonNull V getOrDefault(final @NonNull K key);
 }
