+++ src/main/java/net/kyori/registry/impl/registry/RegistryImpl.java
@@ -26,7 +26,7 @@ package net.kyori.registry.impl.registry;
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
 import com.google.common.collect.Iterators;
-import net.kyori.registry.api.registry.Registry;
+import net.kyori.registry.Registry;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
