+++ src/test/java/net/kyori/registry/map/IncrementalIdMapImplTest.java
@@ -25,27 +25,26 @@ package net.kyori.registry.map;
 
 import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
 import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap;
-import net.kyori.registry.api.map.IncrementalIdMap;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class IncrementalIdMapImplTest {
-    private static final int DEFAULT = -1000;
-    private final IncrementalIdMap<String> map = IncrementalIdMap.create(new Int2ObjectOpenHashMap<>(), new Object2IntOpenHashMap<String>() {
-        {
-            this.defaultReturnValue(DEFAULT);
-        }
-    }, value -> value == -1000);
-
-    @Test
-    void testPut() {
-        final int i0 = this.map.put("abc");
-        assertEquals(0, i0);
-        assertEquals("abc", this.map.get(i0));
-        final int i1 = this.map.put("def");
-        assertEquals(1, i1);
-        assertEquals("def", this.map.get(i1));
-        assertEquals("abc", this.map.get(i0));
+  private static final int DEFAULT = -1000;
+  private final IncrementalIdMap<String> map = IncrementalIdMap.create(new Int2ObjectOpenHashMap<>(), new Object2IntOpenHashMap<String>() {
+    {
+      this.defaultReturnValue(DEFAULT);
     }
+  }, value -> value == -1000);
+
+  @Test
+  void testPut() {
+    final int i0 = this.map.put("abc");
+    assertEquals(0, i0);
+    assertEquals("abc", this.map.get(i0));
+    final int i1 = this.map.put("def");
+    assertEquals(1, i1);
+    assertEquals("def", this.map.get(i1));
+    assertEquals("abc", this.map.get(i0));
+  }
 }
