--- src/test/java/net/kyori/registry/registry/unidirectional/DefaultValueUnidirectionalRegistryTest.java
@@ -1,47 +1,0 @@
-/*
- * This file is part of registry, licensed under the MIT License.
- *
- * Copyright (c) 2018-2019 KyoriPowered
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package net.kyori.registry.registry.unidirectional;
-
-import net.kyori.registry.api.registry.Registry;
-import net.kyori.registry.impl.registry.DefaultedRegistryImpl;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class DefaultValueUnidirectionalRegistryTest {
-    @Test
-    void testDefaultKeyValue() {
-        final DefaultedRegistryImpl<String, String> container = new DefaultedRegistryImpl<>(Registry.create(), "_default");
-        assertEquals("_default", container.defaultKey());
-
-        assertNull(container.get("aaa"));
-        assertNull(container.get("ccc"));
-
-        container.register("_default", "bbb");
-        assertEquals("bbb", container.get("aaa"));
-        assertNotNull(container.get("ccc"));
-
-        assertEquals("bbb", container.getOrDefault("ccc"));
-    }
-}
